Fix metadata place
